hvl2wav multichannel hack
This outputs a wav file containing each individual HVL/AHX channel as stereo (for a total of module channel * 2 channels in the output file).
For best results when using this with AHX files, set the -x0 option and ignore every other channel in the output since AHX channels are typically mono.